Move VPP service methods to the correct file
